Move DAO classes to database.dao sub-package.
