+++ src/main/java/cloudgene/mapred/database/dao/UserDao.java
@@ -1,4 +1,4 @@
-package cloudgene.mapred.database;
+package cloudgene.mapred.database.dao;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
