+++ src/main/java/cloudgene/mapred/database/util/Fixtures.java
@@ -5,8 +5,8 @@ import org.slf4j.LoggerFactory;
 
 import cloudgene.mapred.core.Template;
 import cloudgene.mapred.core.User;
-import cloudgene.mapred.database.TemplateDao;
-import cloudgene.mapred.database.UserDao;
+import cloudgene.mapred.database.dao.TemplateDao;
+import cloudgene.mapred.database.dao.UserDao;
 import cloudgene.mapred.util.HashUtil;
 
 public class Fixtures {
