+++ src/main/java/cloudgene/mapred/server/auth/AuthenticationService.java
@@ -12,7 +12,7 @@ import org.reactivestreams.Subscription;
 
 import cloudgene.mapred.core.ApiToken;
 import cloudgene.mapred.core.User;
-import cloudgene.mapred.database.UserDao;
+import cloudgene.mapred.database.dao.UserDao;
 import cloudgene.mapred.server.Application;
 import cloudgene.mapred.server.responses.ValidatedApiTokenResponse;
 import io.micronaut.security.authentication.Authentication;
