+++ src/main/java/cloudgene/mapred/server/auth/DatabaseAuthenticationProvider.java
@@ -8,7 +8,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import cloudgene.mapred.core.User;
-import cloudgene.mapred.database.UserDao;
+import cloudgene.mapred.database.dao.UserDao;
 import cloudgene.mapred.server.Application;
 import cloudgene.mapred.util.HashUtil;
 import io.micronaut.http.HttpRequest;
