+++ src/test/java/cloudgene/mapred/api/v2/users/ApiTokensTest.java
@@ -17,7 +17,7 @@ import org.junit.jupiter.api.TestInstance;
 
 import cloudgene.mapred.TestApplication;
 import cloudgene.mapred.core.User;
-import cloudgene.mapred.database.UserDao;
+import cloudgene.mapred.database.dao.UserDao;
 import cloudgene.mapred.database.util.Database;
 import cloudgene.mapred.jobs.AbstractJob;
 import cloudgene.mapred.util.CloudgeneClientRestAssured;
