+++ src/test/java/cloudgene/mapred/api/v2/users/RegisterUserTest.java
@@ -8,7 +8,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 import cloudgene.mapred.core.User;
-import cloudgene.mapred.database.UserDao;
+import cloudgene.mapred.database.dao.UserDao;
 import cloudgene.mapred.server.services.UserService;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
