+++ src/test/java/cloudgene/mapred/api/v2/users/UpdatePasswordTest.java
@@ -12,7 +12,7 @@ import org.junit.jupiter.api.TestInstance;
 
 import cloudgene.mapred.TestApplication;
 import cloudgene.mapred.core.User;
-import cloudgene.mapred.database.UserDao;
+import cloudgene.mapred.database.dao.UserDao;
 import cloudgene.mapred.database.util.Database;
 import cloudgene.mapred.util.HashUtil;
 import cloudgene.mapred.util.TestMailServer;
