+++ src/test/java/cloudgene/mapred/jobs/PriorityThreadPoolExecutorTest.java
@@ -12,7 +12,7 @@ import org.junit.jupiter.api.Test;
 
 import cloudgene.mapred.TestApplication;
 import cloudgene.mapred.core.User;
-import cloudgene.mapred.database.UserDao;
+import cloudgene.mapred.database.dao.UserDao;
 import cloudgene.mapred.jobs.workspace.WorkspaceFactory;
 import cloudgene.mapred.util.Settings;
 import cloudgene.mapred.wdl.WdlApp;
