+++ src/test/java/cloudgene/mapred/jobs/WrongWorkspaceTest.java
@@ -10,8 +10,8 @@ import org.junit.jupiter.api.Test;
 
 import cloudgene.mapred.TestApplication;
 import cloudgene.mapred.core.User;
-import cloudgene.mapred.database.JobDao;
-import cloudgene.mapred.database.UserDao;
+import cloudgene.mapred.database.dao.JobDao;
+import cloudgene.mapred.database.dao.UserDao;
 import cloudgene.mapred.jobs.workspace.WorkspaceFactory;
 import cloudgene.mapred.util.Settings;
 import cloudgene.mapred.wdl.WdlApp;
